Fix ssh-add calls for machines where there is no ssh-askpass
Debian and derivates (at least) don't install ssh-askpass by default.
So when ssh-add is called without a pty, it tires and fails to call
ssh-askpass.  Use `start-process' with a filter function to catch the
passphrase prompt instead of `call-process' in that case.

Furthermore, if ssh-add fails for some reason (apart from the user
cancelling), be sure to report the error message.